Update Migration files

+++ database/migrations/2024_07_18_185137_create_nplan_day_recipe_ingredients_table.php
@@ -13,20 +13,22 @@ class CreateNplanDayRecipeIngredientsTable extends Migration
      */
     public function up()
     {
-        Schema::create('nplan_day_recipe_ingredients', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('nplan_day_recipe_id');
-            $table->enum('type', ['main', 'sub']);
-            $table->text('name');
-            $table->bigInteger('quantity');
-            $table->enum('unit', ['tsp', 'tbsp', 'fl oz', 'cup', 'pt', 'qt', 'gal', 'ml', 'l', 'oz', 'lb', 'g', 'kg']);
-            $table->bigInteger('scaled_quantity');
-            $table->enum('scaled_unit', ['tsp', 'tbsp', 'fl oz', 'cup', 'pt', 'qt', 'gal', 'ml', 'l', 'oz', 'lb', 'g', 'kg']);
-            $table->timestamps();
-            $table->softDeletes();
+        if (!Schema::hasTable('nplan_day_recipe_ingredients')) {
+            Schema::create('nplan_day_recipe_ingredients', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('nplan_day_recipe_id');
+                $table->enum('type', ['main', 'sub']);
+                $table->text('name');
+                $table->bigInteger('quantity');
+                $table->enum('unit', ['tsp', 'tbsp', 'fl oz', 'cup', 'pt', 'qt', 'gal', 'ml', 'l', 'oz', 'lb', 'g', 'kg']);
+                $table->bigInteger('scaled_quantity');
+                $table->enum('scaled_unit', ['tsp', 'tbsp', 'fl oz', 'cup', 'pt', 'qt', 'gal', 'ml', 'l', 'oz', 'lb', 'g', 'kg']);
+                $table->timestamps();
+                $table->softDeletes();
 
-            $table->foreign('nplan_day_recipe_id')->references('id')->on('nutrition_plan_day_recipes')->onDelete('cascade');
-        });
+                $table->foreign('nplan_day_recipe_id')->references('id')->on('nutrition_plan_day_recipes')->onDelete('cascade');
+            });
+        }
     }
 
     /**
